msm: eva: Reducing the memory footprint of core structure
The "msm_cvp_core" structure is currently allocated using
"kmalloc", which returns memory that is both virtually and
physically contiguous. However, the size of this structure
has grown significantly (~51 KB), requiring approximately
an order-4 allocation (64 KB). Such large contiguous memory
allocations are not always guaranteed to succeed. To address
this, we are moving the allocation of the largest element
within this structure to a "vmalloc" based approach.

Change-Id: If9dd3a5c79bd6a1064171a0dca2e58d23a85ed17
Signed-off-by: Pranay Varma Kopanati <pkopanat@qti.qualcomm.com>